Relocate CLEO Utils
